Set new backend as main backend
