+++ .gitignore
@@ -8,3 +8,4 @@ launch.json
 deputies.json
 db.sqlite3
 last_deputies.json
+__old-backend
