+++ backend/deputies/activity.py
