+++ backend/requirements.txt
@@ -1,32 +1,32 @@
 async-generator==1.10
-attrs==22.1.0
-beautifulsoup4==4.11.1
-certifi==2022.12.7
-charset-normalizer==2.1.1
+attrs==23.1.0
+beautifulsoup4==4.12.2
+certifi==2023.5.7
+charset-normalizer==3.1.0
 click==8.1.3
-discord.py==2.2.2
-exceptiongroup==1.0.4
+discord.py==2.3.1
+exceptiongroup==1.1.1
 Flask==2.3.2
-Flask-Cors==3.0.10
+Flask-Cors==4.0.0
 h11==0.14.0
 idna==3.4
-importlib-metadata==5.0.0
+importlib-metadata==6.7.0
 itsdangerous==2.1.2
 Jinja2==3.1.2
-MarkupSafe==2.1.1
+MarkupSafe==2.1.3
 outcome==1.2.0
 PySocks==1.7.1
 pytz==2022.7
 requests==2.31.0
-selenium==4.6.0
+selenium==4.10.0
 six==1.16.0
 sniffio==1.3.0
 sortedcontainers==2.4.0
 soupsieve==2.3.2.post1
 trio==0.22.0
-trio-websocket==0.9.2
-urllib3==1.26.12
-Werkzeug==2.2.3
+trio-websocket==0.10.3
+urllib3==2.0.3
+Werkzeug==2.3.6
 wsproto==1.2.0
-zipp==3.10.0
-lxml==4.9.1
+zipp==3.15.0
+lxml==4.9.2
