+++ backend/utils/beacon.py
